+++ src/components/App/index.js
@@ -19,6 +19,7 @@ import Research from '../../components/LandingPages/Research'
 import Services from '../../components/LandingPages/Services'
 import Libraries from '../../components/LandingPages/Libraries'
 import About from '../../components/LandingPages/About'
+import Contact from '../LandingPages/Contact'
 import DatabaseList from '../../components/DatabaseList'
 import SubjectList from '../../components/SubjectList'
 import rootReducers from '../../reducers'
@@ -29,11 +30,29 @@ import NotFound from '../../components/Messages/NotFound'
 
 import Config from '../../shared/Configuration'
 
+import { LINK_CLICK } from '../Link'
+import { SET_SEARCH, SAVE_SEARCH_PREFERENCE } from '../../actions/search.js'
+
+const analyticsActions = [LINK_CLICK, SET_SEARCH, SAVE_SEARCH_PREFERENCE]
+
+const analytics = () => next => action => {
+  window.dataLayer = window.dataLayer || []
+  if (analyticsActions.indexOf(action.type) > -1) {
+    window.dataLayer.push({
+      event: action.type,
+      ...action,
+    })
+  }
+
+  return next(action)
+}
+
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 const store = createStore(
   rootReducers,
   composeEnhancers(applyMiddleware(
-    thunkMiddleware // lets us dispatch() functions
+    thunkMiddleware, // lets us dispatch() functions
+    analytics
   ))
 )
 
@@ -51,6 +70,7 @@ const App = (props) => {
               <Route exact path='/chat' component={ChatPage} />
               <Route exact path='/courses' component={Courses} />
               <Route exact path='/hours' component={Hours} />
+              <Route exact path='/contact-us' component={Contact} />
               <Route exact path='/floor/:id' component={ContentfulFloor} />
               <Route exact path='/news/:id' component={ContentfulNews} />
               <Route exact path='/event/:id' component={ContentfulEvent} />