+++ src/tests/components/Link/index.test.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import { shallow } from 'enzyme'
-import LibLink from '../../../components/Link'
+import { LibLink } from '../../../components/Link'
 import { Link } from 'react-router-dom'
 
 const setup = (props) => {