+++ src/components/App/index.js
@@ -1,5 +1,5 @@
 import React from 'react'
-import { Route, Switch } from 'react-router'
+import { Route, Switch, withRouter } from 'react-router'
 import { Provider } from 'react-redux'
 import { createStore, applyMiddleware, compose } from 'redux'
 
@@ -21,6 +21,7 @@ import Research from '../../components/LandingPages/Research'
 import Services from '../../components/LandingPages/Services'
 import Libraries from '../../components/LandingPages/Libraries'
 import About from '../../components/LandingPages/About'
+import Contact from '../LandingPages/Contact'
 import DatabaseList from '../../components/DatabaseList'
 import SubjectList from '../../components/SubjectList'
 import rootReducers from '../../reducers'
@@ -29,47 +30,71 @@ import Rewrite from './Rewrite'
 
 import NotFound from '../../components/Messages/NotFound'
 
+import Config from '../../shared/Configuration'
+
+import { LINK_CLICK } from '../Link'
+import { SET_SEARCH, SAVE_SEARCH_PREFERENCE } from '../../actions/search.js'
+
+const analyticsActions = [LINK_CLICK, SET_SEARCH, SAVE_SEARCH_PREFERENCE]
+
+const analytics = () => next => action => {
+  window.dataLayer = window.dataLayer || []
+  if (analyticsActions.indexOf(action.type) > -1) {
+    window.dataLayer.push({
+      event: action.type,
+      ...action,
+    })
+  }
+
+  return next(action)
+}
+
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose
 const store = createStore(
   rootReducers,
   composeEnhancers(applyMiddleware(
-    thunkMiddleware // lets us dispatch() functions
+    thunkMiddleware, // lets us dispatch() functions
+    analytics
   ))
 )
 
-const App = () => {
+const App = (props) => {
   return (
-    <Provider store={store}>
-      <div>
-        <Rewrite />
-        <PageWrapper>
-          <Switch>
-            <Route exact path='/' component={Home} />
-            <Route exact path='/about' component={About} />
-            <Route exact path='/chat' component={ChatPage} />
-            <Route exact path='/courses' component={Courses} />
-            <Route exact path='/hours' component={Hours} />
-            <Route exact path='/events' component={Events} />
-            <Route exact path='/news' component={News} />
-            <Route exact path='/floor/:id' component={ContentfulFloor} />
-            <Route exact path='/news/:id' component={ContentfulNews} />
-            <Route exact path='/event/:id' component={ContentfulEvent} />
-            <Route exact path='/libraries' component={Libraries} />
-            <Route exact path='/items-requests' component={PersonalInfo} />
-            <Route exact path='/subjects' component={SubjectList} />
-            <Route exact path='/database/:id' component={DatabasePage} />
-            <Route exact path='/databases/:id' component={DatabaseList} />
-            <Route exact path='/research' component={Research} />
-            <Route exact path='/secure/:id' component={SecureContentfulPage} />
-            <Route exact path='/services' component={Services} />
-            <Route exact path='/:id' component={ContentfulPage} />
+    <Switch>
+      { Rewrite(props) }
+      <Provider store={store}>
+        <div>
+          <meta id='nd-version' content={Config.version} />
+          <PageWrapper>
+            <Switch>
+              <Route exact path='/' component={Home} />
+              <Route exact path='/about' component={About} />
+              <Route exact path='/chat' component={ChatPage} />
+              <Route exact path='/courses' component={Courses} />
+              <Route exact path='/hours' component={Hours} />
+              <Route exact path='/events' component={Events} />
+              <Route exact path='/news' component={News} />
+              <Route exact path='/contact-us' component={Contact} />
+              <Route exact path='/floor/:id' component={ContentfulFloor} />
+              <Route exact path='/news/:id' component={ContentfulNews} />
+              <Route exact path='/event/:id' component={ContentfulEvent} />
+              <Route exact path='/libraries' component={Libraries} />
+              <Route exact path='/items-requests' component={PersonalInfo} />
+              <Route exact path='/subjects' component={SubjectList} />
+              <Route exact path='/database/:id' component={DatabasePage} />
+              <Route exact path='/databases/:id' component={DatabaseList} />
+              <Route exact path='/research' component={Research} />
+              <Route exact path='/secure/:id' component={SecureContentfulPage} />
+              <Route exact path='/services' component={Services} />
+              <Route exact path='/:id' component={ContentfulPage} />
 
-            <Route path='*' component={NotFound} />
-          </Switch>
-        </PageWrapper>
-      </div>
-    </Provider>
+              <Route path='*' component={NotFound} />
+            </Switch>
+          </PageWrapper>
+        </div>
+      </Provider>
+    </Switch>
   )
 }
 
-export default App
+export default withRouter(App)